Extracted out TipOfTheDay class
